+++ src/mongotq/interface.py
@@ -1,9 +1,7 @@
 import logging
 from typing import Union, List
 
-from pymongo import MongoClient
-
-from .task_queue import TaskQueue
+from mongotq.task_queue import TaskQueue
 
 
 def get_task_queue(database_name: str,
@@ -30,16 +28,3 @@ def get_task_queue(database_name: str,
         queue.create_index()
         logging.info(f'TaskQueue collection created: {collection_name}')
     return queue
-
-
-def get_mongo_client(mongo_host: Union[str, List[str]]) -> MongoClient:
-    """
-    Returns a MongoDB client instance.
-
-    :param mongo_host: MongoDB instance hostname or a qualified URI.
-                       Multiple hosts can be provided as a list.
-    :return: the MongoDB client instance
-    """
-    return MongoClient(host=mongo_host,
-                       tlsAllowInvalidCertificates=True,
-                       readPreference='secondaryPreferred')
